requirements.txt
fastapi
uvicorn[standard]
matplotlib
numpy
pandas
seaborn
scipy
transformers
requests
Pillow
pytesseract
python-dateutil
pythainlp
yfinance
PyMuPDF
torch
torchaudio
torchvision
protobuf
tiktoken
sentencepiece
python-multipart
scikit-learn
python-dotenv